refactor: streamline job result parsing and rendering by consolidating utility functions and removing deprecated code
